Revert "修改接口"
This reverts commit 6ec39f6e4b48010738042c2f64ef8289385f6b10.

+++ app/Http/Controllers/API/ApplicationController.php
@@ -53,27 +53,27 @@ class ApplicationController extends Controller
     {
         $user = Auth::user();
         $application = Notification::find($id);
-        return $this->acceptApplication($application, $user);
+        return $this->acceptApplication($application);
     }
 
     public function reject(RejectApplication $request, $id)
     {
         $user = Auth::user();
         $application = Notification::find($id);
-        return $this->rejectApplication($application, $user);
+        return $this->rejectApplication($application);
     }
 
-    protected function acceptApplication($application, $user)
+    protected function acceptApplication($application)
     {
         $data = json_decode($application->data);
 
         switch ($data->resource_type) {
             case 'Organization':
-                return $this->acceptOrganizationApplication($application, $user);
+                return $this->acceptOrganizationApplication($application);
                 break;
 
             case 'Lecture':
-                return $this->acceptLectureApplication($application, $user);
+                return $this->acceptLectureApplication($application);
                 break;
             
             default:
@@ -82,17 +82,17 @@ class ApplicationController extends Controller
         }
     }
 
-    protected function rejectApplication($application, $user)
+    protected function rejectApplication($application)
     {
         $data = json_decode($application->data);
 
         switch ($data->resource_type) {
             case 'Organization':
-                return $this->rejectOrganizationApplication($application, $user);
+                return $this->rejectOrganizationApplication($application);
                 break;
 
             case 'Lecture':
-                return $this->rejectLectureApplication($application, $user);
+                return $this->acceptLectureApplication($application);
                 break;
             
             default:
@@ -101,12 +101,16 @@ class ApplicationController extends Controller
         }
     }
 
-    public function acceptOrganizationApplication($application, $user)
+    public function acceptOrganizationApplication($application)
     {
         $data = json_decode($application->data);
 
+        $user = Auth::user();
         $organization = Organization::findOrFail($data->resource_id);
         // 检测当前用户的权限
+        if ($organization->creator_id != $user->id) {
+            return response()->json(['error'=>'This action is unauthorized.'], 403);
+        }
         
         // 将用户加入组织
         $users = User::findOrFail($application->notifiable_id);
@@ -116,11 +120,16 @@ class ApplicationController extends Controller
         $application->delete();
     }
 
-    public function acceptLectureApplication($application, $user)
+    public function acceptLectureApplication($application)
     {
         $data = json_decode($application->data);
 
+        $user = Auth::user();
         $lectrue = Lecture::findOrFail($data->resource_id);
+        // 检测当前用户的权限
+        if ($lectrue->user_id != $user->id) {
+            return response()->json(['error'=>'This action is unauthorized.'], 403);
+        }
         
         // 将用户加入课程
         $users = User::findOrFail($application->notifiable_id);
@@ -130,25 +139,19 @@ class ApplicationController extends Controller
         $application->delete();
     }
 
-    public function rejectOrganizationApplication($application, $user)
+    public function rejectOrganizationApplication($application)
     {
         $data = json_decode($application->data);
 
+        $user = Auth::user();
         $organization = Organization::findOrFail($data->resource_id);
+        // 检测当前用户的权限
+        if ($organization->creator_id != $user->id) {
+            return response()->json(['error'=>'This action is unauthorized.'], 403);
+        }
         
         // 发送通知
         $user->notify(new SystemNotification((object)['to' => $application->notifiable_id, 'data' => "加入组织 {$organization->name} 被拒绝"]));
-        $application->delete();
-    }
-
-    public function rejectLectureApplication($application, $user)
-    {
-        $data = json_decode($application->data);
-
-        $lecture = Lecture::findOrFail($data->resource_id);
-        
-        // 发送通知
-        $user->notify(new SystemNotification((object)['to' => $application->notifiable_id, 'data' => "加入授课 {$organization->name} 被拒绝"]));
         $application->delete();
     }
 }
